backend/task_manager_api/requirements.txt
asgiref             
Django              
djangorestframework
pip                 
pytz                
setuptools          
sqlparse            